update version; minor refactoring
